Update C and C++ languages
This updates the C language to C11 and the C++ language to C++17.  It
introduces a new c_preprocessor.lang, because C and C++ have diverged
a bit, and it was difficult to make them continue to share so much.

This also impacted d.lang.

Here I chose to make "final" and "override" keywords, even though they
are context-dependent.  I also chose not to make the alterative
operator spellings into keywords.